requirements-mac.txt
# macOS pip requirements (used by environment-mac.yml)
# Python 3.10 recommended

absl-py==2.3.1
astunparse==1.6.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.61.0
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
iminuit==2.32.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pandas==1.5.3
pillow==12.0.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.15.3
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
tensorflow-estimator==2.10.0
# Removed strict pin for Apple Silicon/Python 3.10 compatibility.
# If needed, pip will resolve a compatible tensorflow-io-gcs-filesystem version.
termcolor==3.2.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.4
wrapt==2.0.1
